v2.1.7  * SF 247 - 248 - 261 - 264 - 265 - 266 - 267  * 67:	protect uploadify library => different file protection added  * protect Downloadfile.php  * SF228: reset personal saltkey (purge personal items)  * SF262: copy of item is in log  * old password in log was badly encoded  * item copy from search page corrected  * some rights checks added before action  * email send to new created user

+++ install/install_ajax.php
@@ -129,7 +129,7 @@ if ( isset($_POST['type']) ){
                 CREATE TABLE IF NOT EXISTS `".$_SESSION['tbl_prefix']."log_items` (
                   `id_item` int(8) NOT NULL,
                   `date` varchar(50) NOT NULL,
-                  `id_user` tinyint(4) NOT NULL,
+                  `id_user` int(8) NOT NULL,
                   `action` varchar(250) NOT NULL,
                   `raison` text NOT NULL
                 ) CHARSET=utf8;");
